Creating GUI layout
Still creating th GUI layout adding more screens such as a title, player
select, and disclaimer scenes. Currently there is at least one more scene
that will have to be completed, and the board scene will need to be updated
before starting work on the model. In addition the title page might get
updated so that it switches out the image instead of creating a new
scene for hints and the disclaimer.